fix: isolate vote request validation
