+++ next.config.js
@@ -23,7 +23,9 @@ const nextConfig = {
     imageSizes: [16, 32, 48, 64, 96, 128, 256, 384],
   },
   experimental: {
-    optimizeCss: true,
+    optimizeCss: {
+      beasties: true,
+    },
   },
   async headers() {
     return [
