+++ package.json
@@ -13,6 +13,7 @@
   "dependencies": {
     "class-variance-authority": "^0.7.1",
     "clsx": "^2.1.1",
+    "beasties": "^0.1.0",
     "lucide-react": "^0.536.0",
     "next": "^15.3.2",
     "react": "^18.3.1",
